+++ src/internal.ts
@@ -11,7 +11,7 @@
  * @param n The value to check.
  * @returns true if the value is numeric, false otherwise.
  */
-function isNumeric(n: unknown): n is number {
+function isNum(n: unknown): n is number {
   return !Number.isNaN(Number(n) - Number.parseFloat(n as string));
 }
 
@@ -20,14 +20,27 @@ function isNumeric(n: unknown): n is number {
  *
  * This function is a type guard that checks if the given value is an array.
  *
- * @param a The value to check.
+ * @param v The value to check.
  * @returns true if the value is an array, false otherwise.
  */
-function isArray(a: unknown): a is unknown[] {
-  return a?.constructor === Array;
+function isArr(v: unknown): v is unknown[] {
+  return v?.constructor === Array;
+}
+
+/**
+ * Checks if the given value is a string.
+ *
+ * This function is a type guard that checks if the given value is a string.
+ *
+ * @param v The value to check.
+ * @returns true if the value is a string, false otherwise.
+ */
+function isStr(v: unknown): v is string {
+  return typeof v === "string";
 }
 
 export {
-  isNumeric,
-  isArray,
+  isNum,
+  isArr,
+  isStr,
 };